153929 Bake IDE AMI - Windows Updates pause

+++ scripts/dot-createlicense.ps1
@@ -13,27 +13,35 @@ function CreateLicence {
 	   [string]$registryValue
    )
 
-   $mypwd = ConvertTo-SecureString -String $password -Force –AsPlainText
-   Import-PfxCertificate –FilePath $licenseFile cert:\\localMachine\\my -Password $mypwd
+   # Check if license file is available to be installed
+   if ( Test-Path $licenseFile )
+   {
 
-   #####################################################################################
-   # Save private key filename & current Machine Guid to registry
-   #####################################################################################
-   $getCert = Get-ChildItem  -path "Cert:\LocalMachine\My" -DNSName $dnsName
+       $mypwd = ConvertTo-SecureString -String $password -Force –AsPlainText
+       Import-PfxCertificate –FilePath $licenseFile cert:\\localMachine\\my -Password $mypwd
 
-   $Thumbprint = $getCert.Thumbprint
+       #####################################################################################
+       # Save private key filename & current Machine Guid to registry
+       #####################################################################################
+       $getCert = Get-ChildItem  -path "Cert:\LocalMachine\My" -DNSName $dnsName
 
-   $keyName=(((Get-ChildItem Cert:\LocalMachine\My | Where-Object {$_.Thumbprint -like $Thumbprint}).PrivateKey).CspKeyContainerInfo).UniqueKeyContainerName
+       $Thumbprint = $getCert.Thumbprint
 
-   New-Item -Path HKLM:\Software\LANSA
-   New-ItemProperty -Path HKLM:\Software\LANSA  -Name $registryValue -PropertyType String -Value $keyName -Force
+       $keyName=(((Get-ChildItem Cert:\LocalMachine\My | Where-Object {$_.Thumbprint -like $Thumbprint}).PrivateKey).CspKeyContainerInfo).UniqueKeyContainerName
 
-   $MachineGuid = Get-ItemProperty -Path HKLM:\SOFTWARE\Microsoft\Cryptography  -Name MachineGuid
-   New-ItemProperty -Path HKLM:\Software\LANSA  -Name PriorMachineGuid -PropertyType String -Value $MachineGuid.MachineGuid -force 
+       New-Item -Path HKLM:\Software\LANSA
+       New-ItemProperty -Path HKLM:\Software\LANSA  -Name $registryValue -PropertyType String -Value $keyName -Force
 
-   # Write any old junk into the license file to obliterate the contents from the disk so it cannot be recovered
-   Get-Process | Out-File $licenseFile
-   #Now delete it from Explorer
-   Remove-Item $licenseFile
+       $MachineGuid = Get-ItemProperty -Path HKLM:\SOFTWARE\Microsoft\Cryptography  -Name MachineGuid
+       New-ItemProperty -Path HKLM:\Software\LANSA  -Name PriorMachineGuid -PropertyType String -Value $MachineGuid.MachineGuid -force 
 
+       # Write any old junk into the license file to obliterate the contents from the disk so it cannot be recovered
+       Get-Process | Out-File $licenseFile
+       #Now delete it from Explorer
+       Remove-Item $licenseFile
+    }
+    else
+    {
+        Write-Output "$LicenseFile does not exist. Presume its already installed."
+    }
 }
